+++ product/interfaces/graphql/service.go
@@ -42,9 +42,10 @@ func (*Service) Models() map[string]config.TypeMapEntry {
 		"Commerce_ProductAttributes": graphql.ModelMapEntry{
 			Type: domain.Attributes{},
 			Fields: map[string]string{
-				"getAttribute":     "Attribute",
-				"getAttributes":    "Attributes",
-				"getAttributeKeys": "AttributeKeys",
+				"getAttribute":       "Attribute",
+				"getAttributes":      "Attributes",
+				"getAttributeKeys":   "AttributeKeys",
+				"getAttributesByKey": "AttributesByKey",
 			},
 		},
 		"Commerce_ProductAttribute":        domain.Attribute{},