+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     - removed depricated viewdata (CartTotals)
 - products:
     - product category breadcrumb is not filled in controller - if you want a breadcrum you can use category data functions
-    - product category fields are changed to use a categoryTeaser
+    - product category fields are changed to use a categoryTeaser    
 - category:
     - Tree object uses a Tree Entity now which contains NOT all category properties. You have to fetch the category details separate on demand:
         - search for usages of the data funcs - they may need changes in rendering the data: `data('category´´..`
@@ -121,3 +121,7 @@ the cart has been adjusted and written back to cache
 **price**
 * IsZero() now uses LikelyEqual() instead of Equal() to avoid issues occurring due to floating-point arithmetic
 
+**product**
+* product attributes:
+  * Added `AttributesByKey` domain method to filter attributes by key and exposed this method as `getAttributesByKey` in GraphQL
+  * GraphQL: Exposing `codeLabel` property in the `Commerce_ProductAttribute` type